update: user profile refactor, layout update
