+++ src/constants/Global.ts
@@ -1,4 +1,4 @@
 // Url example: 'http://10.15.3.5:3000/'
-const Url = 'http://10.15.6.6:3000';
+const Url = 'http://10.19.222.87:3000';
 
 export { Url };
