+++ src/pages/profile/MatchingHistory.tsx
@@ -1,32 +1,14 @@
 import styled from 'styled-components';
 
-import { historyProps } from './ProfileProps';
-
-const RecordWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 1em;
-  justify-content: center;
-  align-items: center;
-  background: rgba( 0, 0, 0, 0 );
-`;
-
-const RecordP = styled.p`
-  border-width: 1px;
-  border-style: solid;
-  white-space: pre-line;
-  text-align: center;
-  background: rgba( 0, 0, 0, 0 );
-`;
+import { historyProps } from './UserProps';
 
 const RecordContainer = styled.div`
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
   width: 70vw;
-  border-width: 2px;
-  border-style: solid;
+  border-width: 1px;
+  border-style: solid hidden;
   display: grid;
   justify-content: center;
+  font-size: 1.5rem;
   grid-template-columns: 1fr 2fr 1fr 2fr;
   grid-template-rows: 1fr 2fr;
   grid-template-areas:
@@ -34,38 +16,47 @@ const RecordContainer = styled.div`
   "result my score opp";
   & > * {
     display: flex;
-    border-width: 1px;
-    border-style: solid;
+    border-width: 0.5px;
+    border-style: solid hidden;
+    border-style: solid solid solid hidden;
     white-space: pre-line;
     align-items: center;
     justify-content: center;
-    background: rgba( 0, 0, 0, 0 );
-    font-family: SBAggroM;
+    background: rgba(0, 0, 0, 0);
+    font-family: SBAggroL;
   }
 `;
 
+const Img = styled.img`
+  display: flex;
+  flex-direction: column;
+  border-radius: 50%;
+  width: 1em;
+  height: 1em;
+  margin: 0.25rem;
+  background: rgba( 0, 0, 0, 0 );
+`;
+
 const Time = styled.div`
+  font-size: 1rem;
   grid-area: time;
 `
 
 const My = styled.div`
   grid-area: my;
-  font-size: 3rem;
 `
 
 const Score = styled.div`
   grid-area: score;
-  font-size: 1rem;
 `
 
 const Opp = styled.div`
   grid-area: opp;
-  font-size: 1rem;
+  border-style: solid hidden;
 `
 
 const Result = styled.div`
   grid-area: result;
-  font-size: 2rem;
 `
 
 const MatchingHistory = ({name, image, history} : {name: string, image: string, history: historyProps}) => {
@@ -73,15 +64,19 @@ const MatchingHistory = ({name, image, history} : {name: string, image: string, 
   const winlose: string = history.win? "win!" : "lose...";
 
   return (
-    <RecordWrapper>
-      <RecordContainer>
-        <Time>{history.played_at}</Time>
-        <Result>{winlose}</Result>
-        <My>{name}</My>
-        <Score>{score}</Score>
-        <Opp>{history.opponent_name}</Opp>
-      </RecordContainer>
-    </RecordWrapper>
+    <RecordContainer>
+      <Time>{history.played_at}</Time>
+      <Result>{winlose}</Result>
+      <My>
+        <Img src={image} alt="not fount" />
+        {name}
+      </My>
+      <Score>{score}</Score>
+      <Opp>
+        <Img src={image} alt="not fount" />
+        {history.opponent_name}
+      </Opp>
+    </RecordContainer>
   );
 }
