+++ src/pages/profile/UserStats.tsx
@@ -1,13 +1,16 @@
-import { useState } from "react";
 import styled from "styled-components";
-import { statsProps } from "./ProfileProps";
+import { statsProps } from "./UserProps";
 
-const Stat = styled.div`
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-  font-size: 4em;
+const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
+  align-items: center;
+  grid-area: stats;
+`
+
+const Stat = styled.div`
+  margin: 0.5em;
+  font-size: 2em;
   justify-content: center;
   align-items: center;
   background: rgba( 0, 0, 0, 0 );
@@ -15,11 +18,28 @@ const Stat = styled.div`
 
 const UserStats = ({stats} : {stats : statsProps}) => {
   return (
-    <>
-    <Stat>{"Rank: " + stats.ladder_score}</Stat>
-    <Stat>{"Arcade: " + stats.arcade_score}</Stat>
-    <Stat>{"Win: " + stats.wins.toString() + " Lose: " + stats.loses.toString()}</Stat>
-    </>
+    <Wrapper>
+      <table>
+        <tbody>
+          <tr>
+            <td>
+              <Stat>{"Rank: " + stats.ladder_score}</Stat>
+            </td>
+            <td>
+              <Stat>{"Arcade: " + stats.arcade_score}</Stat>
+            </td>
+          </tr>
+          <tr>
+            <td>
+              <Stat>{"Win: " + stats.wins.toString()}</Stat>
+            </td>
+            <td>
+              <Stat>{"Lose: " + stats.loses.toString()}</Stat>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </Wrapper>
   );
 }
